+++ docs/changelog.md
@@ -1261,7 +1261,7 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 
 0.5.1: fix: add PG to futures list
 
-0.5.2: fix: air_zhenqi.py rename air_city_dict to air_city_list
+0.5.2: fix: air_zhenqi.py rename air_city_dict to air_city_table
 
 0.5.3: add: add two fields into covid_163
 
@@ -3143,4 +3143,6 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.4.43 fix: fix stock_fhps_em interface
 
 1.4.44 fix: fix stock_dxsyl_em interface
+
+1.4.45 fix: fix air_quality_rank interface
 ```